Medical_final_project_github_folder/requirements.txt
streamlit==1.56.0
chromadb==1.5.7
groq==1.1.2
Pillow==12.2.0
audio-recorder-streamlit==0.0.10
opencv-python-headless==4.13.0.92
numpy==2.1.0
pytesseract==0.3.13
pdfplumber==0.11.9
pdf2image==1.17.0
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
